wip: abstract common fns required for consensus
